requirements.txt
# Requirements
python-dotenv
requests
matplotlib
Pillow